Live recording + add track button
